// src/codegen/refractedlabs/flowtrade/flow.ts
export enum FlowStatus {
  /** WAITING - the flow is waiting to be started */
  WAITING = 0,
  /** ACTIVE - the flow is active */
  ACTIVE = 1,
  /** ENDED - the flow has been ended. this is a final state */
  ENDED = 2,
  /** STOPPED - the flow has been stopped. this is a final state */
  STOPPED = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface Flow {
  /** auto-generated identifier for a flow */
  id: bigint;
  /** the address of flow creator and owner */
  creator: string;
  /** informational data about the flow */
  flowInfo: FlowInfo;
  /** the time that swap distribution starts */
  startTime: Timestamp;
  /** the time that flow is ended */
  endTime: Timestamp;
  /**
   * the interval in which the distribution index is updated and hence tokens are swapped
   * if dist_interval is 0, the flow is updated every time in or out tokens are increased or decreased
   * if dist_interval is equal to the duration of flow, it means that all of the tokens are swapped once after the flow ends
   */
  distInterval: Duration;
  /** the address that the swapped tokens are sent to after the flow ends */
  treasuryAddress: string;
  /** total amount creator provided to be swapped */
  totalTokenOut: Coin;
  /** the accepted token to buy the out tokens */
  tokenInDenom: string;
  /** the time from then the buyers can claim their purchased tokens */
  tokenOutClaimableAfter: Timestamp;
  /** the time from then the flow creator can claim the swapped tokens */
  tokenInClaimableAfter: Timestamp;
  /** whether the flow can be stopped by the flow's creator */
  stoppable: boolean;
  /** whether to allow buyers to claim their tokens immediately after the flow is stopped */
  allowImmediateTokenOutClaimIfStopped: boolean;
  /** whether to allow flow's creator to claim tokens immediately after the flow is stopped */
  allowImmediateTokenInClaimIfStopped: boolean;
  /** a global index for the amount of purchase that has already been applied */
  distIndex: string;
  /** the last time the dist index is updated */
  lastDistUpdate: Timestamp;
  /** the amount of remaining out tokens to sell */
  tokenOutBalance: string;
  /** the current amount of token-in provided to buy token-out */
  tokenInBalance: string;
  /** the amount of already spent in tokens */
  spentTokenIn: string;
  /** the total number of users shares */
  totalShares: string;
  /** the last swap price */
  livePrice: string;
  /** the current status of flow */
  status: FlowStatus;
  /** the amount of creation deposit that should be revenue to the flow creator */
  creationDeposit?: Coin;
  /** the fee ratio taken from token-out, this is copied from module params at the time of flow creation */
  tokenOutFeeRatio: string;
  /** the fee ratio taken from token-in, this is copied from module params at the time of flow creation */
  tokenInFeeRatio: string;
  /**
   * if true, the swapped tokens are automatically sent to the treasury address when the flow ends
   * this option is only available when another module creates a flow using keeper api
   */
  automaticTreasuryCollection: boolean;
  /** the amount of spent in tokens that have been claimed by the flow's creator */
  claimedTokenIn: string;
  /** whether the flow is checked out, meaning the creation deposit and any remaining out tokens are transferred back to the creator */
  checkedOut: boolean;
}
/** Flow holds information and price calculations for a flow trade */
export interface FlowSDKType {
  id: bigint;
  creator: string;
  flow_info: FlowInfoSDKType;
  start_time: TimestampSDKType;
  endTime: TimestampSDKType;
  dist_interval: DurationSDKType;
  treasury_address: string;
  total_token_out: CoinSDKType;
  token_in_denom: string;
  token_out_claimable_after: TimestampSDKType;
  token_in_claimable_after: TimestampSDKType;
  stoppable: boolean;
  allow_immediate_token_out_claim_if_stopped: boolean;
  allow_immediate_token_in_claim_if_stopped: boolean;
  dist_index: string;
  last_dist_update: TimestampSDKType;
  token_out_balance: string;
  token_in_balance: string;
  spent_token_in: string;
  total_shares: string;
  live_price: string;
  status: FlowStatus;
  creation_deposit?: CoinSDKType;
  token_out_fee_ratio: string;
  token_in_fee_ratio: string;
  automatic_treasury_collection: boolean;
  claimed_token_in: string;
  checked_out: boolean;
}
/** Informational data about the flow */
export interface FlowInfo {
  /** a name of the flow */
  name: string;
  /** a short description about the flow */
  description: string;
  /** a referring url */
  url: string;
}
/** Informational data about the flow */
export interface FlowInfoSDKType {
  name: string;
  description: string;
  url: string;
}
// ...
export const FlowInfo = {
  encode(message: FlowInfo, writer: BinaryWriter = BinaryWriter.create()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    if (message.description !== "") {
      writer.uint32(18).string(message.description);
    }
    if (message.url !== "") {
      writer.uint32(26).string(message.url);
    }
    return writer;
  },
  decode(input: BinaryReader | Uint8Array, length?: number): FlowInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFlowInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.name = reader.string();
          break;
        case 2:
          message.description = reader.string();
          break;
        case 3:
          message.url = reader.string();
          break;
        default:
          reader.skipType(tag & 7);
          break;
      }
    }
    return message;
  },
  fromJSON(object: any): FlowInfo {
    return {
      name: isSet(object.name) ? String(object.name) : "",
      description: isSet(object.description) ? String(object.description) : "",
      url: isSet(object.url) ? String(object.url) : ""
    };
  },
  toJSON(message: FlowInfo): unknown {
    const obj: any = {};
    message.name !== undefined && (obj.name = message.name);
    message.description !== undefined && (obj.description = message.description);
    message.url !== undefined && (obj.url = message.url);
    return obj;
  },
  fromPartial(object: Partial<FlowInfo>): FlowInfo {
    const message = createBaseFlowInfo();
    message.name = object.name ?? "";
    message.description = object.description ?? "";
    message.url = object.url ?? "";
    return message;
  }
};